# chore: collect relationships so they can be batch inserted