Baseline:

```
2024-04-26T14:07:32.672525Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: parse: 2s 492ms 106us 342ns    
2024-04-26T14:07:32.672536Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: ingest 1: 2ms 564us 988ns    
2024-04-26T14:07:32.672546Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: ingest 2: 12s 226ms 606us 90ns    
2024-04-26T14:07:32.672555Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: commit: 385us 427ns    
2024-04-26T14:07:32.672569Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: query: 4ms 289us 159ns    
test graph::sbom::spdx::tests::perf::ingest_spdx_medium ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 43 filtered out; finished in 16.23s
```

With this PR:

```
2024-04-26T16:11:14.768068Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: parse: 2s 491ms 140us 782ns    
2024-04-26T16:11:14.768077Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: ingest 1: 2ms 426us 564ns    
2024-04-26T16:11:14.768085Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: ingest 2: 8s 561ms 340us 8ns    
2024-04-26T16:11:14.768092Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: commit: 347us 146ns    
2024-04-26T16:11:14.768099Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: query: 3ms 557us 136ns    
test graph::sbom::spdx::tests::perf::ingest_spdx_medium ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 43 filtered out; finished in 12.56s
```